examples/langgraph-memanto-pr472/requirements.txt
langgraph>=0.3.0
langgraph-cli>=0.1.0
langchain-core>=0.3.0
langchain-community>=0.3.0
memanto>=0.1.0
moorcheh-sdk>=1.3.5
python-dotenv>=1.0.0
httpx>=0.25.0
